dictionary supplement scripts moved to a new blueprint, special buttons added to admin dictionaries
