--- get_eng_words.py
@@ -1,30 +1,0 @@
-from datetime import datetime
-
-from english_dict_generator import eng_dict_generator
-from webapp import create_app
-from webapp.db import db
-from webapp.dictionary.models import EnglishWord
-from webapp.dictionary.get_transcription import get_transcription
-
-app = create_app()
-
-
-def save_words_in_db(words_dict):
-    for word in words_dict:
-        word_exist = EnglishWord.query.filter(EnglishWord.word_itself == word).count()
-        if not word_exist:
-            new_word = EnglishWord(
-                word_itself=word,
-                translation_rus=words_dict[word],
-                transcription=get_transcription(word),
-                imported_time=datetime.now()
-                )
-            db.session.add(new_word)
-    db.session.commit()
-
-
-words_dict = eng_dict_generator()
-
-if __name__ == "__main__":
-    with app.app_context():
-        save_words_in_db(words_dict)
