+++ webapp/templates/dictionary/engdict_index.html
@@ -22,7 +22,19 @@
                                         </div>
                                     </div>
                                 </form>
-                                <p>Всего английских слов: {{ english_list_len }}</p>
+                                <div class='row'>
+                                    <div class='col-6'>
+                                        <p>Всего английских слов: {{ english_list_len }}</p>
+                                    </div>
+                                    <div class='col-6' align='right'>
+                                        <a href="{{ url_for(
+                                        'dict_supplement_and_parsers.admin_engdict_supplement',
+                                        username=current_user.username) }}"
+                                        class="btn btn-outline-info">
+                                            Пополнить английский словарь словами из текстового файла
+                                        </a>
+                                    </div>
+                                </div>
                                 <table class="table table-bordered">
                                     <thead>
                                         <tr>
